% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toastr_clear.R
\name{toastr_clear_all}
\alias{toastr_clear_all}
\title{Remove existing toastr notifications}
\usage{
toastr_clear_all(with_animation = FALSE)
}
\arguments{
\item{with_animation}{If \code{FALSE}, toasts are destroyed instantly. Else they disappear according to their \code{hideMethod}.}
}
\description{
Remove existing toastr notifications
}

