% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinytoastr.R
\name{useToastr}
\alias{useToastr}
\title{Initialize the toastr notification engine}
\usage{
useToastr()
}
\value{
The HTML tags to put into the \code{<head>} of the HTML file.
}
\description{
Call this function once, from the top of your Shiny UI definition.
Here is an example: \preformatted{  ui <- shinyUI(fluidPage(
    useToastr(),
    pageWithSidebar(
      headerPanel("Header"),
      sidebarPanel(
        ...
      ),
      mainPanel(
        ...
      )
    )
  ))
}
}
\examples{
## See above
}
\seealso{
\code{\link{toastr_success}}, \code{\link{toastr_info}},
  \code{\link{toastr_warning}}, \code{\link{toastr_error}}
}

